Use local_inner_macros to resolve all local macros within $crate
This fixes the following error when using Rust 2018 style macro imports.

    use log::info;

    fn main() {
        info!("...");
    }

    error: cannot find macro `log!` in this scope
     --> src/main.rs:4:5
      |
    4 |     info!("...");
      |     ^^^^^^^^^^^^^

The `local_inner_macros` modifier resolves all macro invocations made
from within that macro as macros in the same crate. So if `info!`
expands to an invocation of `log!` then this would be resolved as
`$crate::log!` rather than requiring the caller to have `log` in scope.

The attribute is ignored by pre-2018 compilers so log will continue to
work as normal with #[macro_use].

In the future when dropping compatibility with pre-2018 compilers we can
remove the `local_inner_macros` modifier and use our own explicit
`$crate::` prefixes on invocations of local macros.