Improve error when missing macro argument
The log! macro matches part of its arguments as $($arg:tt)+, but the other
macros match it as $($arg:tt)*. That means if a user calls error!() or any of
the other macros with an empty argument, it successfully matches the macro rules
of the first macro they called but then turns into an invalid call to log!,
leading to a confusing error message.

This PR uses $($arg:tt)+ consistently to match arguments.

Example:

    use log::error;

    fn main() {
        error!();
    }

Before:

    error: expected identifier, found `,`
     --> src/main.rs:4:5
      |
    4 |     error!();
      |     ^^^^^^^^^
      |     |
      |     expected identifier
      |     help: remove this comma
      |
      = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)

After:

    error: unexpected end of macro invocation
     --> src/main.rs:4:5
      |
    4 |     error!();
      |     ^^^^^^^^^ missing tokens in macro arguments