lib: Implement AtomicUsize for platforms without atomics
Some platforms don't provide a AtomicUsize. Instead of just failing to
build with this error:

291 | use std::sync::atomic::{AtomicUsize, Ordering};
    |                         ^^^^^^^^^^^ no `AtomicUsize` in `sync::atomic`

let's instead add a fake AtomicUsize.

As the platform doesn't have atomics we can't implement an Atomic
version of AtomicUsize, so this version is not atomic. Any platform
without atomics is unlikely to have multiple cores, so this shouldn't be
a problem.

This is somewhat based on:
https://github.com/japaric/heapless/commit/940d2e9d7f7ff57f9922b012fcad54cd46b673ae

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>